Refactor FileFilter to encapsulate gitignore setup and remove the ignore module
